move ITeam and IEntity to global type file because of using in many files
